Use relative path mirror instead of github for litehtml submodule
To avoid github clone issues, we should clone from a url relative to
the parent module, which will resolve either to code.qt.io or gerrit
when cloned from either of those.

This is safe to do, now that the litehtml repo is mirrored to both
gerrit and code.qt.io.

Task-number: QTQAINFRA-6817
Change-Id: Ic0ceee1bfe4986ac5fa116c88ab9779973993b88
Reviewed-by: Toni Saario <toni.saario@qt.io>